ENG-959: Update docs to encourage using explorer for creating collections (#102)
* Update get-started.mdx

Added this video to show them how to create a simple collection using Explorer to encourage users to use Explorer.

* Update get-started.mdx

I have taken it off from Getting Started as you mentioned. Will now add it over to the Collections and add a new pull request for the same.

* Update collections.mdx

Moved over the SDK Collections code to the Collections section. Let me know if there are any more changes here to be made.

* Created a new heading for Creating Collections Programmatically

* Made the necessary changes as you had mentioned.

* Made a small change to the doc

* Split sections on creating collection

---------

Co-authored-by: Shaki <112879608+shaki150697@users.noreply.github.com>
Co-authored-by: Shaki150697 <shaki@polybase.xyz>